allow manual seal to produce non-empty blocks with BABE
To enable the creation of non-empty blocks in manual mode with the BABE runtime, we need this workaround. BABE generates 2 digests (PreRuntime and Seal), whereas manual seal generates only one (PreRuntime). If we attempt to create a block containing only 1 user extrinsic (a common scenario in our integration tests), this control would otherwise raise an error, hindering the functionality of our tests.